added draft map.
Added a draft map to display the sampling by county from a geojson that
is produced by the package. I chose not to go with the leaflet ajax
package because the examples on leaflets website do not use that
strategy. I will write the geojson to a .js file and give it a variable
name there. The I will source the file in the map.html leaflet section.