Moved popup text formatting to R side.
All features will have a popup text. Instead of formatting the text in
javascript then it make more sense to keep the map.html a little more
consistent, always called 'popup_text' from the geojson file as the
popup text on a feature. Then that text should be formatted on the R
side in a variable of that name.