Points outside polygons
Added a warning for points that lie outside of the polygon layer when
the matching is done. Also corrected the lan and kommune data so that
the attached dataframe is not factor data but character data. Now the
matching function returns a list with polygons and the points that were
not matched.