Added dynamic legend
Added the feature that the legend changes depending on the input
data. This allows the map to remain the same even after positives have
been added. Also removed all positives from the map in the datacleaning
stage. We will need to discuss the release of each new positive point
and make a pressd release prior to allowing the points to be published
on this map.