Responsive elements
Added the height of the map device to be responsive to the height of the
viewport, which I hope will be interpreted as the iframe. This is in
order to allow the map to adjust to size of the responsive iframe on a
mobile browser.

Added the maxBounds argument to the map function to allow the map to
'bounce back' when the user scrolls out of bounds