Moved back to state without dependancies
Some of the dependancies of the packge were causing the package not to
build. Not exactly sure why, but this commit removes almost everything
that is not currently being used in production and then we can start to
add stuff back as we need it. The last build of the master branch
resulted in the following failure on travis:

Warning message:
packages ‘xts,’, ‘leaflet,’, ‘mapview,’, ‘htmlwidgets,’ are not available (for R version 3.3.1)

If this commit passes in the dev-T branch I will pull it into the master