Added html TABLE
Added a function to build an html table using xtable. I think that an
extension of this will become the addition of the data.table javascript
library to make the tables interactive. This will require adding a line
to the head calling the data.table.js and .css; and adding the call to
the data.table function also in the <head>. Then the class of the
<table> needs the addition of the second data.table class.