Added <thead> and <tbody> to table
These elements of the table are used by data.table as well as a table
identifier. I also added a table id to allow data.table to reference
it. Interestingly xtable does not use this form <thead> <tbody> in its
tables where the kable function in the knitr package does. I would
consider added knitr instead and then migrate to th ekable function to
write html but so far this fix will work.