Travis update
The documentation for Rttravis indicates that the tool scans the
DESCRIPTION files to determien which packages to install. Therefore I
removed the packages listed in the DESCRIPTION file from the the
r_packages.

Also a call to a specific version of R. Not sure if this is the correct
nomenclature as the documentation was a little difficult to understand.