BUG FIX
I had accidentally commented out the line that filters the reported
cases by those that were approved to be published. Therefore on Dec22 at
6:05 a map was published with incorrect information. A positive point on
Gotland and a positive point outside Stockholm was published. In the
data these were not confirmed to be HighPathAI and therefore should
have been published even if they were approved. The error was detected
at 7:30 and correct at 8:15.

A possible addition to the code could be to add a filter on the Aivh5330
variable, to only allow those with the word "HÖGPATOGEN" but this may be
a low sensitivity test since the field is free text. I will stick with
just the approved list for now and assume that I will be able to keep up
with the information flow of new highpath cases.