correction to wget to force authorization on first attempt.
wget does not supply username or password untill challenged by the
server to do so. In this case it fails twice, the first time by
no-auth, the second because it is reusing the previous communications
protocol but with auth. Then the thrird time it increases the
communicatio protocol and uses auth and it
works. "--auth-no-challenge" ensures that it doesn't do this testing.