Removed comments from print.xtable
The xtable package inserts a comment and timestamp in each table that
it writes with the print.xtable() function. I previously thought that
this would be inert, however, since the timestamp changes daily it
causes the .json data to have unnecessary git diffs when the source
data has not changed.

With this removal I expect that it will be easier to see the relevant
daily data diffs if a problem needs to be debugged.