Added date calculation to legend
The influenza map has a legend that describes the date of the oldest
sample and the date that the map was updated. This is 180 days prior
to the map update date. This implementation is no really in line with
the vision of simlicity on the js side and doing to calculations on
the R side. I should probably just calculate the startdate in R and
then pass it to js in the .json file. For now this will work. It also
allowed me to learn a little about date and string handling in js.