Another example fo a time series
getting closer to the total functionality neccessary to build time
series

I will next split the writting of the timeseries data to json format
and the options that are dataset specific to two functions. Followed
by the matching options that may be needed in when writting the html
code in order to have multiple axes in the plot.