Order of bars and colors
The dataseries are plotted in reverse order to the order they appear
in the json data. This means if you want the lines on top of the bars
you must put the lines first in the data.

Also transparency in hex colours doesn't seem to work in IE therefore
I dropped that.